src/lib/ai/content-scheduler.ts
import { OpenAI } from 'openai';
import { enhancedCache } from '../enhanced-cache';
import { LinkBioProfile } from '@/types/link-bio';

const openai = new OpenAI({
  apiKey: process.env.OPENAI_API_KEY,
});

interface TimeSlot {
  day: string;
  hour: number;
  score: number;
  reason: string;
}

interface ContentSchedule {
  recommendedTimes: TimeSlot[];
  frequency: string;
  timezone: string;
  audienceInsights: {
    primaryTimezones: string[];
    peakActivityHours: number[];
    demographicFactors: Record<string, number>;
  };
}

interface ContentPlan {
  schedule: ContentSchedule;
  contentVariations: {
    timeSlot: TimeSlot;
    content: string;
    targetAudience: string;
    expectedEngagement: number;
  }[];
}

export class ContentScheduler {
  private static instance: ContentScheduler;

  private constructor() {}

  static getInstance(): ContentScheduler {
    if (!ContentScheduler.instance) {
      ContentScheduler.instance = new ContentScheduler();
    }
    return ContentScheduler.instance;
  }

  async generateOptimalSchedule(
    profile: LinkBioProfile,
    analyticsData: any,
    preferences: {
      timezone?: string;
      excludedDays?: string[];
      minPostsPerWeek?: number;
      maxPostsPerWeek?: number;
    } = {}
  ): Promise<ContentSchedule> {
    const cacheKey = `schedule:${profile.id}:${JSON.stringify(preferences)}`;
    const cached = await enhancedCache.get(cacheKey);
    if (cached) return cached;

    const completion = await openai.chat.completions.create({
      model: "gpt-4",
      messages: [
        {
          role: "system",
          content: "Analyze user data and generate optimal content schedule.",
        },
        {
          role: "user",
          content: `Generate optimal posting schedule for:
                   Profile: ${JSON.stringify(profile)}
                   Analytics: ${JSON.stringify(analyticsData)}
                   Preferences: ${JSON.stringify(preferences)}`,
        },
      ],
      temperature: 0.7,
    });

    const schedule = this.parseSchedule(completion.choices[0].message.content);
    await enhancedCache.set(cacheKey, schedule, 3600 * 24); // Cache for 24 hours
    return schedule;
  }

  async generateContentPlan(
    profile: LinkBioProfile,
    schedule: ContentSchedule,
    contentTemplate: string
  ): Promise<ContentPlan> {
    const variations = await Promise.all(
      schedule.recommendedTimes.map(async timeSlot => {
        const variation = await this.generateContentVariation(
          contentTemplate,
          timeSlot,
          profile
        );
        return {
          timeSlot,
          ...variation,
        };
      })
    );

    return {
      schedule,
      contentVariations: variations,
    };
  }

  async analyzeAudienceEngagement(
    profile: LinkBioProfile,
    analyticsData: any
  ): Promise<{
    bestTimes: TimeSlot[];
    audienceSegments: Record<string, number>;
    engagementPatterns: Record<string, any>;
  }> {
    const completion = await openai.chat.completions.create({
      model: "gpt-4",
      messages: [
        {
          role: "system",
          content: "Analyze audience engagement patterns and provide insights.",
        },
        {
          role: "user",
          content: `Analyze engagement for:
                   Profile: ${JSON.stringify(profile)}
                   Analytics: ${JSON.stringify(analyticsData)}`,
        },
      ],
      temperature: 0.7,
    });

    return this.parseEngagementAnalysis(completion.choices[0].message.content);
  }

  async predictEngagement(
    content: string,
    timeSlot: TimeSlot,
    historicalData: any
  ): Promise<{
    score: number;
    factors: Record<string, number>;
    recommendations: string[];
  }> {
    const completion = await openai.chat.completions.create({
      model: "gpt-4",
      messages: [
        {
          role: "system",
          content: "Predict content engagement based on historical data and timing.",
        },
        {
          role: "user",
          content: `Predict engagement for:
                   Content: ${content}
                   TimeSlot: ${JSON.stringify(timeSlot)}
                   Historical Data: ${JSON.stringify(historicalData)}`,
        },
      ],
      temperature: 0.7,
    });

    return this.parseEngagementPrediction(completion.choices[0].message.content);
  }

  async optimizePostTiming(
    schedule: ContentSchedule,
    realTimeData: any
  ): Promise<ContentSchedule> {
    // Adjust schedule based on real-time engagement data
    const adjustedSchedule = { ...schedule };
    
    // Implement real-time optimization logic
    adjustedSchedule.recommendedTimes = schedule.recommendedTimes.map(slot => ({
      ...slot,
      score: this.calculateRealTimeScore(slot, realTimeData),
    }));

    return adjustedSchedule;
  }

  private async generateContentVariation(
    template: string,
    timeSlot: TimeSlot,
    profile: LinkBioProfile
  ): Promise<{
    content: string;
    targetAudience: string;
    expectedEngagement: number;
  }> {
    const completion = await openai.chat.completions.create({
      model: "gpt-4",
      messages: [
        {
          role: "system",
          content: "Generate content variation optimized for specific time slot.",
        },
        {
          role: "user",
          content: `Generate variation for:
                   Template: ${template}
                   TimeSlot: ${JSON.stringify(timeSlot)}
                   Profile: ${JSON.stringify(profile)}`,
        },
      ],
      temperature: 0.8,
    });

    return this.parseContentVariation(completion.choices[0].message.content);
  }

  private calculateRealTimeScore(slot: TimeSlot, realTimeData: any): number {
    // Implementation of real-time score calculation
    return 0;
  }

  private parseSchedule(content: string | null): ContentSchedule {
    if (!content) {
      return {
        recommendedTimes: [],
        frequency: 'daily',
        timezone: 'UTC',
        audienceInsights: {
          primaryTimezones: [],
          peakActivityHours: [],
          demographicFactors: {},
        },
      };
    }
    
    // Implementation of schedule parsing
    return {
      recommendedTimes: [],
      frequency: 'daily',
      timezone: 'UTC',
      audienceInsights: {
        primaryTimezones: [],
        peakActivityHours: [],
        demographicFactors: {},
      },
    };
  }

  private parseEngagementAnalysis(content: string | null): any {
    if (!content) return {};
    
    // Implementation of engagement analysis parsing
    return {};
  }

  private parseEngagementPrediction(content: string | null): any {
    if (!content) return {};
    
    // Implementation of engagement prediction parsing
    return {};
  }

  private parseContentVariation(content: string | null): any {
    if (!content) return {};
    
    // Implementation of content variation parsing
    return {};
  }
}

export const contentScheduler = ContentScheduler.getInstance();
